backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
requests
smtplib
email-validator
